+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2026-03-24
+
+### :toolbox: Added
+
+- Add Format-HtmlCell cmdlet to the module to format HTML table cells with specified background color, text color, and font size
+- Add Set-ImageOpacity cmdlet to the module to set the opacity of an image file
+- Add new parameters to Add-NodeImage cmdlet to set image opacity for node images
+- Add WaterMark support to SVG output format
+  - Use the DiaConvertImageToPDF c# net4.8 package to add watermark support to SVG output format
+
+### :arrows_clockwise: Changed
+
+- Update module version to v1.0.5
+- Update edge length validation range, and help message to reflect the new range of 0 to 50 in Add-NodeEdge function
+- Update Pester workflow to run tests and upload results with code coverage support
+- Optimize aspect ratio calculation in Get-BestImageAspectRatio function
+
+### :wrench: Fixed
+
+- Fix pester tests failing in Windows Pwsh 5.1+
+
 ## [1.0.4] - 2026-03-17
 
 ### :arrows_clockwise: Changed